requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
